refactor: code cleanup + adding inline comments
